feat: navigator 구현 (#7)
