+++ lib/main.dart
@@ -1,287 +1,24 @@
 import 'package:flutter/material.dart';
-import 'package:fluttertoast/fluttertoast.dart';
+import 'package:minions/Home.dart';
+import 'package:minions/ScreenA.dart';
+import 'package:minions/ScreenB.dart';
 
-void main() => runApp(const MyApp());
+void main() => runApp(MyApp());
 
 class MyApp extends StatelessWidget {
   const MyApp({super.key});
 
   @override
   Widget build(BuildContext context) {
-    return const MaterialApp(
+    return MaterialApp(
       title: "Despicable Me",
-      home: MyPage(),
+      initialRoute: '/',
+      routes: {
+        '/': (context) => Home(),
+        '/a': (context) => ScreenA(),
+        '/b': (context) => ScreenB()
+      },
       debugShowCheckedModeBanner: false,
-    );
-  }
-}
-
-class MyPage extends StatelessWidget {
-  const MyPage({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      backgroundColor: Colors.amber[700],
-      appBar: AppBar(
-        title: const Text('MINIONS'),
-        backgroundColor: Colors.amber[800],
-        centerTitle: true,
-        elevation: 0.0,
-        actions: [
-          IconButton(
-            icon: Icon(Icons.shopping_cart),
-            onPressed: () {
-              print('Shopping cart button button is clicked');
-            },
-          ),
-          IconButton(
-            icon: Icon(Icons.search),
-            onPressed: () {
-              print('Search button is clicked');
-            },
-          ),
-        ],
-      ),
-      drawer: Drawer(
-        child: ListView(
-          padding: EdgeInsets.zero,
-          children: [
-            UserAccountsDrawerHeader(
-              currentAccountPicture: CircleAvatar(
-                backgroundImage: AssetImage('assets/kevin.png'),
-                backgroundColor: Colors.white,
-              ),
-              otherAccountsPictures: [
-                CircleAvatar(
-                  backgroundColor: Colors.white,
-                  backgroundImage: AssetImage('assets/agnes.png'),
-                ),
-              ],
-              accountName: Text('Bob'),
-              accountEmail: Text('slz6k@naver.com'),
-              onDetailsPressed: () {
-                print('Arrow is clicked');
-              },
-              decoration: BoxDecoration(
-                  color: Colors.amber[500],
-                  borderRadius: BorderRadius.only(
-                      bottomLeft: Radius.circular(40.0),
-                      bottomRight: Radius.circular(40.0))),
-            ),
-            ListTile(
-              leading: Icon(
-                Icons.home,
-                color: Colors.grey[850],
-              ),
-              title: Text('Home'),
-              onTap: () {
-                print('Home is clicked');
-              },
-              trailing: Icon(Icons.add),
-            ),
-            ListTile(
-              leading: Icon(
-                Icons.settings,
-                color: Colors.grey[850],
-              ),
-              title: Text('Setting'),
-              onTap: () {
-                print('Setting is clicked');
-              },
-              trailing: Icon(Icons.add),
-            ),
-            ListTile(
-              leading: Icon(
-                Icons.question_answer,
-                color: Colors.grey[850],
-              ),
-              title: Text('Q&A'),
-              onTap: () {
-                print('Q&A is clicked');
-              },
-              trailing: Icon(Icons.add),
-            )
-          ],
-        ),
-      ),
-      body: Padding(
-        padding: EdgeInsets.fromLTRB(30.0, 40.0, 0.0, 0.0),
-        child: Column(
-          crossAxisAlignment: CrossAxisAlignment.start,
-          children: [
-            Center(
-              child: CircleAvatar(
-                backgroundImage: AssetImage('assets/alert_minion.gif'),
-                radius: 60.0,
-              ),
-            ),
-            Divider(
-              height: 60.0,
-              color: Colors.grey[850],
-              thickness: 0.5,
-              endIndent: 30.0,
-            ),
-            Text(
-              'NAME',
-              style: TextStyle(
-                color: Colors.white,
-                letterSpacing: 2.0,
-              ),
-            ),
-            SizedBox(
-              height: 10.0,
-            ),
-            Text(
-              'MINIONS',
-              style: TextStyle(
-                  color: Colors.white,
-                  letterSpacing: 2.0,
-                  fontSize: 28.0,
-                  fontWeight: FontWeight.bold),
-            ),
-            SizedBox(
-              height: 30.0,
-            ),
-            Text(
-              'MINI LEVEL',
-              style: TextStyle(
-                color: Colors.white,
-                letterSpacing: 2.0,
-              ),
-            ),
-            SizedBox(
-              height: 10.0,
-            ),
-            Text(
-              '14',
-              style: TextStyle(
-                  color: Colors.white,
-                  letterSpacing: 2.0,
-                  fontSize: 28.0,
-                  fontWeight: FontWeight.bold),
-            ),
-            SizedBox(
-              height: 30.0,
-            ),
-            Row(
-              children: [
-                Icon(Icons.check_circle_outline),
-                SizedBox(
-                  width: 10.0,
-                ),
-                Text(
-                  'Kevin',
-                  style: TextStyle(
-                    fontSize: 16.0,
-                    letterSpacing: 1.0,
-                  ),
-                ),
-              ],
-            ),
-            Row(
-              children: [
-                Icon(Icons.check_circle_outline),
-                SizedBox(
-                  width: 10.0,
-                ),
-                Text(
-                  'Stuart',
-                  style: TextStyle(
-                    fontSize: 16.0,
-                    letterSpacing: 1.0,
-                  ),
-                ),
-              ],
-            ),
-            Row(
-              children: [
-                Icon(Icons.check_circle_outline),
-                SizedBox(
-                  width: 10.0,
-                ),
-                Text(
-                  'Bob',
-                  style: TextStyle(
-                    fontSize: 16.0,
-                    letterSpacing: 1.0,
-                  ),
-                ),
-              ],
-            ),
-            Center(
-              child: CircleAvatar(
-                backgroundImage: AssetImage('assets/minions.png'),
-                radius: 40.0,
-                backgroundColor: Colors.amber[700],
-              ),
-            ),
-            MySnackBar(),
-            Center(
-              child: TextButton(
-                onPressed: () {
-                  flutterToast();
-                },
-                child: Text(
-                  'Toast',
-                  style: TextStyle(color: Colors.white),
-                ),
-                style: ButtonStyle(
-                    backgroundColor: MaterialStateProperty.all(Colors.amber)),
-              ),
-            )
-          ],
-        ),
-      ),
-    );
-  }
-}
-
-void flutterToast() {
-  Fluttertoast.showToast(
-    msg: 'Minions',
-    gravity: ToastGravity.CENTER,
-    backgroundColor: Colors.amberAccent,
-    fontSize: 20.0,
-    textColor: Colors.white,
-    toastLength: Toast.LENGTH_SHORT,
-  );
-}
-
-class MySnackBar extends StatelessWidget {
-  const MySnackBar({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return Center(
-      child: TextButton(
-        style: ButtonStyle(
-            backgroundColor: MaterialStateProperty.all(Colors.amber)),
-        onPressed: () {
-          ScaffoldMessenger.of(context).showSnackBar(
-            SnackBar(
-              action: SnackBarAction(
-                label: 'Action',
-                textColor: Colors.grey[850],
-                onPressed: () {
-                  print('SnackBar is clicked');
-                },
-              ),
-              content: const Text(
-                'Hellow',
-                textAlign: TextAlign.center,
-                style: TextStyle(color: Colors.white),
-              ),
-              backgroundColor: Colors.amber,
-              duration: Duration(microseconds: 3000),
-            ),
-          );
-        },
-        child: const Text(
-          'Show me',
-          style: TextStyle(color: Colors.white),
-        ),
-      ),
     );
   }
 }
